WIP: Adding command line wrapper
Branch: feature/add-command-line-wrapper
The application failed due to an issue in the raw command line
execution, something to which the application is heavilly based. Much
propbabily, this error was not where it was initially tested, but have
no idea why failed in the current system where I stumbled this problem.
So, a command line wrapper would be appropriated to deal with different
environments's idiosyncrasies where I need to run raw command line
executions.